fix: utility function `getLineItemsGroupedByVendor`
