+++ client-app/core/utilities/index.ts
@@ -2,6 +2,7 @@ export * from "./address";
 export * from "./categories";
 export * from "./common";
 export * from "./facets";
+export * from "./line-items";
 export * from "./logger";
 export * from "./menu";
 export * from "./product";
