+++ client-app/shared/account/types/index.ts
@@ -1,5 +1,3 @@
-import { OrderLineItemType, Vendor } from "@/xapi";
-
 export enum RegistrationKind {
   personal = "personal",
   organization = "organization",
@@ -62,6 +60,3 @@ export type OrdersFilterChipsItem = {
   value: string | Date;
   label: string;
 };
-
-export type TGroupItem = { items: OrderLineItemType[]; vendor?: Vendor };
-export type TGroupedItems = Record<string, TGroupItem>;
